refactored Synth and Poly/monophonic classes
